Add protocol support to staker and validator test helps. Add tests for legacy add_stake.
