+++ primitives/account/tests/staking_contract/add_stake_policy.rs
@@ -1,0 +1,906 @@
+use nimiq_account::*;
+use nimiq_database::{mdbx::MdbxDatabase, traits::Database};
+use nimiq_primitives::{coin::Coin, policy::Policy, transaction::TransactionError};
+use nimiq_test_log::test;
+use nimiq_transaction::account::staking_contract::IncomingStakingTransactionData;
+
+use super::*;
+
+fn add_stake(protocol_version: u16) {
+    let env = MdbxDatabase::new_volatile(Default::default()).unwrap();
+    let accounts = Accounts::new(env.clone());
+    let data_store = accounts.data_store(&Policy::STAKING_CONTRACT_ADDRESS);
+    let block_state = BlockState::new(2, 2, protocol_version);
+    let mut db_txn = env.write_transaction();
+    let mut db_txn = (&mut db_txn).into();
+
+    let (validator_address, staker_address, mut staking_contract) =
+        make_sample_contract(data_store.write(&mut db_txn), Some(150_000_000));
+    let staker_address = staker_address.unwrap();
+    let staker_keypair = ed25519_key_pair(STAKER_PRIVATE_KEY);
+
+    // Works in the valid case.
+    let tx = make_signed_incoming_transaction(
+        IncomingStakingTransactionData::AddStake {
+            staker_address: staker_address.clone(),
+        },
+        150_000_000,
+        &staker_keypair,
+    );
+
+    let mut tx_logger = TransactionLog::empty();
+    let receipt = staking_contract
+        .commit_incoming_transaction(
+            &tx,
+            &block_state,
+            data_store.write(&mut db_txn),
+            &mut tx_logger,
+        )
+        .expect("Failed to commit transaction");
+
+    assert_eq!(
+        receipt,
+        Some(
+            AddStakeReceipt {
+                credited_balance: BalanceType::Active
+            }
+            .into()
+        )
+    );
+    assert_eq!(
+        tx_logger.logs,
+        vec![Log::Stake {
+            staker_address: staker_address.clone(),
+            validator_address: Some(validator_address.clone()),
+            value: tx.value,
+            credited_balance: BalanceType::Active,
+        }]
+    );
+
+    let staker = staking_contract
+        .get_staker(&data_store.read(&db_txn), &staker_address)
+        .expect("Staker should exist");
+
+    assert_eq!(staker.address, staker_address);
+    assert_eq!(staker.active_balance, Coin::from_u64_unchecked(300_000_000));
+    assert_eq!(staker.delegation, Some(validator_address.clone()));
+
+    let validator = staking_contract
+        .get_validator(&data_store.read(&db_txn), &validator_address)
+        .expect("Validator should exist");
+
+    assert_eq!(
+        validator.total_stake,
+        Coin::from_u64_unchecked(Policy::VALIDATOR_DEPOSIT + 300_000_000)
+    );
+    assert_eq!(validator.num_stakers, 1);
+
+    assert_eq!(
+        staking_contract.balance,
+        Coin::from_u64_unchecked(Policy::VALIDATOR_DEPOSIT + 300_000_000)
+    );
+
+    assert_eq!(
+        staking_contract.active_validators.get(&validator_address),
+        Some(&Coin::from_u64_unchecked(
+            Policy::VALIDATOR_DEPOSIT + 300_000_000
+        ))
+    );
+
+    // Revert the transaction.
+    let mut tx_logger = TransactionLog::empty();
+    staking_contract
+        .revert_incoming_transaction(
+            &tx,
+            &block_state,
+            receipt,
+            data_store.write(&mut db_txn),
+            &mut tx_logger,
+        )
+        .expect("Failed to revert transaction");
+
+    assert_eq!(
+        tx_logger.logs,
+        vec![Log::Stake {
+            staker_address: staker_address.clone(),
+            validator_address: Some(validator_address.clone()),
+            value: tx.value,
+            credited_balance: BalanceType::Active,
+        }]
+    );
+
+    let staker = staking_contract
+        .get_staker(&data_store.read(&db_txn), &staker_address)
+        .expect("Staker should exist");
+
+    assert_eq!(staker.address, staker_address);
+    assert_eq!(staker.active_balance, Coin::from_u64_unchecked(150_000_000));
+    assert_eq!(staker.delegation, Some(validator_address.clone()));
+
+    let validator = staking_contract
+        .get_validator(&data_store.read(&db_txn), &validator_address)
+        .expect("Validator should exist");
+
+    assert_eq!(
+        validator.total_stake,
+        Coin::from_u64_unchecked(Policy::VALIDATOR_DEPOSIT + 150_000_000)
+    );
+    assert_eq!(validator.num_stakers, 1);
+
+    assert_eq!(
+        staking_contract.balance,
+        Coin::from_u64_unchecked(Policy::VALIDATOR_DEPOSIT + 150_000_000)
+    );
+
+    assert_eq!(
+        staking_contract.active_validators.get(&validator_address),
+        Some(&Coin::from_u64_unchecked(
+            Policy::VALIDATOR_DEPOSIT + 150_000_000
+        ))
+    );
+}
+
+fn add_stake_policy_priority_to_active_balance(protocol_version: u16) {
+    // -----------------------------------
+    // Test setup:
+    // -----------------------------------
+    let mut staker_setup = StakerSetup::setup_staker_with_inactive_retired_balance(
+        ValidatorState::Active,
+        Policy::MINIMUM_STAKE,
+        Policy::MINIMUM_STAKE + 1,
+        50_000_000,
+        protocol_version,
+    );
+    assert!(staker_setup.active_stake < staker_setup.retired_stake);
+    assert!(staker_setup.active_stake < staker_setup.inactive_stake);
+    let data_store = staker_setup
+        .accounts
+        .data_store(&Policy::STAKING_CONTRACT_ADDRESS);
+    let mut db_txn = staker_setup.env.write_transaction();
+    let mut db_txn = (&mut db_txn).into();
+    let staker_keypair = ed25519_key_pair(STAKER_PRIVATE_KEY);
+
+    // -----------------------------------
+    // Test execution:
+    // -----------------------------------
+    // Add stake operation credits to active stake.
+    let tx = make_signed_incoming_transaction(
+        IncomingStakingTransactionData::AddStake {
+            staker_address: staker_setup.staker_address.clone(),
+        },
+        Policy::MINIMUM_STAKE,
+        &staker_keypair,
+    );
+
+    let mut tx_logs = TransactionLog::empty();
+    let receipt = staker_setup
+        .staking_contract
+        .commit_incoming_transaction(
+            &tx,
+            &staker_setup.before_release_block_state,
+            data_store.write(&mut db_txn),
+            &mut tx_logs,
+        )
+        .expect("Failed to commit transaction");
+
+    assert_eq!(
+        receipt,
+        Some(
+            AddStakeReceipt {
+                credited_balance: BalanceType::Active
+            }
+            .into()
+        )
+    );
+
+    assert_eq!(
+        tx_logs.logs,
+        vec![Log::Stake {
+            staker_address: staker_setup.staker_address.clone(),
+            validator_address: Some(staker_setup.validator_address.clone()),
+            value: Coin::from_u64_unchecked(Policy::MINIMUM_STAKE),
+            credited_balance: BalanceType::Active,
+        }]
+    );
+
+    let staker = staker_setup
+        .staking_contract
+        .get_staker(&data_store.read(&db_txn), &staker_setup.staker_address)
+        .expect("Staker should exist");
+
+    assert_eq!(
+        staker.delegation,
+        Some(staker_setup.validator_address.clone())
+    );
+    assert_eq!(
+        staker.active_balance,
+        Coin::from_u64_unchecked(Policy::MINIMUM_STAKE * 2)
+    );
+    assert_eq!(
+        staker.inactive_balance,
+        Coin::from_u64_unchecked(Policy::MINIMUM_STAKE + 1)
+    );
+    assert_eq!(staker.inactive_from, Some(328));
+    assert_eq!(
+        staker.retired_balance,
+        Coin::from_u64_unchecked(Policy::MINIMUM_STAKE * 5)
+    );
+
+    let validator = staker_setup
+        .staking_contract
+        .get_validator(&data_store.read(&db_txn), &staker_setup.validator_address)
+        .unwrap();
+
+    assert_eq!(validator.num_stakers, 1);
+    assert_eq!(
+        validator.total_stake,
+        Coin::from_u64_unchecked(Policy::VALIDATOR_DEPOSIT + Policy::MINIMUM_STAKE * 2)
+    );
+
+    // Reverts correctly.
+    staker_setup
+        .staking_contract
+        .revert_incoming_transaction(
+            &tx,
+            &staker_setup.before_release_block_state,
+            receipt,
+            data_store.write(&mut db_txn),
+            &mut TransactionLog::empty(),
+        )
+        .expect("Failed to commit transaction");
+
+    let staker = staker_setup
+        .staking_contract
+        .get_staker(&data_store.read(&db_txn), &staker_setup.staker_address)
+        .expect("Staker should exist");
+    assert_eq!(
+        staker.delegation,
+        Some(staker_setup.validator_address.clone())
+    );
+
+    assert_eq!(staker.active_balance, staker_setup.active_stake);
+    assert_eq!(staker.inactive_balance, staker_setup.inactive_stake);
+    assert_eq!(
+        staker.inactive_from,
+        Some(staker_setup.effective_block_state.number)
+    );
+    assert_eq!(staker.retired_balance, staker_setup.retired_stake);
+
+    let validator = staker_setup
+        .staking_contract
+        .get_validator(&data_store.read(&db_txn), &staker_setup.validator_address)
+        .unwrap();
+    assert_eq!(validator.num_stakers, 1);
+    assert_eq!(
+        validator.total_stake,
+        Coin::from_u64_unchecked(Policy::VALIDATOR_DEPOSIT + Policy::MINIMUM_STAKE)
+    );
+}
+
+#[test]
+fn add_stake_works() {
+    add_stake(Policy::max_supported_version());
+}
+
+#[test]
+fn add_stake_policy_priority_to_active_balance_works() {
+    add_stake_policy_priority_to_active_balance(Policy::max_supported_version());
+}
+
+/// Adding stake in absence of active balance should prioritize the maximum between the inactive and retired balance.
+#[test]
+fn add_stake_policy_maximum_between_inactive_retired(protocol_version: u16) {
+    // -----------------------------------
+    // Test setup:
+    // -----------------------------------
+    let mut staker_setup = StakerSetup::setup_staker_with_inactive_retired_balance(
+        ValidatorState::Active,
+        0,
+        Policy::MINIMUM_STAKE + 1,
+        50_000_000,
+        Policy::max_supported_version(),
+    );
+    assert!(staker_setup.active_stake < staker_setup.retired_stake);
+    assert!(staker_setup.active_stake < staker_setup.inactive_stake);
+    let data_store = staker_setup
+        .accounts
+        .data_store(&Policy::STAKING_CONTRACT_ADDRESS);
+    let mut db_txn = staker_setup.env.write_transaction();
+    let mut db_txn = (&mut db_txn).into();
+    let staker_keypair = ed25519_key_pair(STAKER_PRIVATE_KEY);
+
+    // -----------------------------------
+    // Test execution:
+    // -----------------------------------
+    // Add stake operation credits to active stake.
+    let tx = make_signed_incoming_transaction(
+        IncomingStakingTransactionData::AddStake {
+            staker_address: staker_setup.staker_address.clone(),
+        },
+        Policy::MINIMUM_STAKE,
+        &staker_keypair,
+    );
+
+    let mut tx_logs = TransactionLog::empty();
+    let receipt = staker_setup
+        .staking_contract
+        .commit_incoming_transaction(
+            &tx,
+            &staker_setup.before_release_block_state,
+            data_store.write(&mut db_txn),
+            &mut tx_logs,
+        )
+        .expect("Failed to commit transaction");
+
+    assert_eq!(
+        receipt,
+        Some(
+            AddStakeReceipt {
+                credited_balance: BalanceType::Retired
+            }
+            .into()
+        )
+    );
+
+    assert_eq!(
+        tx_logs.logs,
+        vec![Log::Stake {
+            staker_address: staker_setup.staker_address.clone(),
+            validator_address: Some(staker_setup.validator_address.clone()),
+            value: Coin::from_u64_unchecked(Policy::MINIMUM_STAKE),
+            credited_balance: BalanceType::Retired,
+        }]
+    );
+
+    let staker = staker_setup
+        .staking_contract
+        .get_staker(&data_store.read(&db_txn), &staker_setup.staker_address)
+        .expect("Staker should exist");
+
+    assert_eq!(
+        staker.delegation,
+        Some(staker_setup.validator_address.clone())
+    );
+    assert_eq!(staker.active_balance, Coin::ZERO);
+    assert_eq!(
+        staker.inactive_balance,
+        Coin::from_u64_unchecked(Policy::MINIMUM_STAKE + 1)
+    );
+    assert_eq!(staker.inactive_from, Some(328));
+    assert_eq!(
+        staker.retired_balance,
+        Coin::from_u64_unchecked(Policy::MINIMUM_STAKE * 6)
+    );
+
+    let validator = staker_setup
+        .staking_contract
+        .get_validator(&data_store.read(&db_txn), &staker_setup.validator_address)
+        .unwrap();
+
+    assert_eq!(validator.num_stakers, 1);
+    assert_eq!(
+        validator.total_stake,
+        Coin::from_u64_unchecked(Policy::VALIDATOR_DEPOSIT)
+    );
+
+    // Reverts correctly.
+    staker_setup
+        .staking_contract
+        .revert_incoming_transaction(
+            &tx,
+            &staker_setup.before_release_block_state,
+            receipt,
+            data_store.write(&mut db_txn),
+            &mut TransactionLog::empty(),
+        )
+        .expect("Failed to commit transaction");
+
+    let staker = staker_setup
+        .staking_contract
+        .get_staker(&data_store.read(&db_txn), &staker_setup.staker_address)
+        .expect("Staker should exist");
+    assert_eq!(
+        staker.delegation,
+        Some(staker_setup.validator_address.clone())
+    );
+
+    assert_eq!(staker.active_balance, staker_setup.active_stake);
+    assert_eq!(staker.inactive_balance, staker_setup.inactive_stake);
+    assert_eq!(
+        staker.inactive_from,
+        Some(staker_setup.effective_block_state.number)
+    );
+    assert_eq!(staker.retired_balance, staker_setup.retired_stake);
+
+    let validator = staker_setup
+        .staking_contract
+        .get_validator(&data_store.read(&db_txn), &staker_setup.validator_address)
+        .unwrap();
+    assert_eq!(validator.num_stakers, 1);
+    assert_eq!(
+        validator.total_stake,
+        Coin::from_u64_unchecked(Policy::VALIDATOR_DEPOSIT)
+    );
+}
+
+/// Adding stake to the inactive stake works.
+#[test]
+fn add_stake_policy_to_inactive(protocol_version: u16) {
+    // -----------------------------------
+    // Test setup:
+    // -----------------------------------
+    let mut staker_setup = StakerSetup::setup_staker_with_inactive_retired_balance(
+        ValidatorState::Active,
+        0,
+        50_000_000,
+        Policy::MINIMUM_STAKE + 1,
+        Policy::max_supported_version(),
+    );
+    assert!(staker_setup.active_stake < staker_setup.retired_stake);
+    assert!(staker_setup.active_stake < staker_setup.inactive_stake);
+    let data_store = staker_setup
+        .accounts
+        .data_store(&Policy::STAKING_CONTRACT_ADDRESS);
+    let mut db_txn = staker_setup.env.write_transaction();
+    let mut db_txn = (&mut db_txn).into();
+    let staker_keypair = ed25519_key_pair(STAKER_PRIVATE_KEY);
+
+    // -----------------------------------
+    // Test execution:
+    // -----------------------------------
+    // Add stake operation credits to active stake.
+    let tx = make_signed_incoming_transaction(
+        IncomingStakingTransactionData::AddStake {
+            staker_address: staker_setup.staker_address.clone(),
+        },
+        Policy::MINIMUM_STAKE,
+        &staker_keypair,
+    );
+
+    let mut tx_logs = TransactionLog::empty();
+    let receipt = staker_setup
+        .staking_contract
+        .commit_incoming_transaction(
+            &tx,
+            &staker_setup.before_release_block_state,
+            data_store.write(&mut db_txn),
+            &mut tx_logs,
+        )
+        .expect("Failed to commit transaction");
+
+    assert_eq!(
+        receipt,
+        Some(
+            AddStakeReceipt {
+                credited_balance: BalanceType::Inactive
+            }
+            .into()
+        )
+    );
+
+    assert_eq!(
+        tx_logs.logs,
+        vec![Log::Stake {
+            staker_address: staker_setup.staker_address.clone(),
+            validator_address: Some(staker_setup.validator_address.clone()),
+            value: Coin::from_u64_unchecked(Policy::MINIMUM_STAKE),
+            credited_balance: BalanceType::Inactive,
+        }]
+    );
+
+    let staker = staker_setup
+        .staking_contract
+        .get_staker(&data_store.read(&db_txn), &staker_setup.staker_address)
+        .expect("Staker should exist");
+    assert_eq!(
+        staker.delegation,
+        Some(staker_setup.validator_address.clone())
+    );
+
+    assert_eq!(staker.active_balance, Coin::ZERO);
+    assert_eq!(
+        staker.inactive_balance,
+        Coin::from_u64_unchecked(Policy::MINIMUM_STAKE * 6)
+    );
+    assert_eq!(staker.inactive_from, Some(328));
+    assert_eq!(
+        staker.retired_balance,
+        Coin::from_u64_unchecked(Policy::MINIMUM_STAKE + 1)
+    );
+
+    let validator = staker_setup
+        .staking_contract
+        .get_validator(&data_store.read(&db_txn), &staker_setup.validator_address)
+        .unwrap();
+
+    assert_eq!(validator.num_stakers, 1);
+    assert_eq!(
+        validator.total_stake,
+        Coin::from_u64_unchecked(Policy::VALIDATOR_DEPOSIT)
+    );
+
+    // Reverts correctly.
+    staker_setup
+        .staking_contract
+        .revert_incoming_transaction(
+            &tx,
+            &staker_setup.before_release_block_state,
+            receipt,
+            data_store.write(&mut db_txn),
+            &mut TransactionLog::empty(),
+        )
+        .expect("Failed to commit transaction");
+
+    let staker = staker_setup
+        .staking_contract
+        .get_staker(&data_store.read(&db_txn), &staker_setup.staker_address)
+        .expect("Staker should exist");
+    assert_eq!(
+        staker.delegation,
+        Some(staker_setup.validator_address.clone())
+    );
+
+    assert_eq!(staker.active_balance, staker_setup.active_stake);
+    assert_eq!(staker.inactive_balance, staker_setup.inactive_stake);
+    assert_eq!(
+        staker.inactive_from,
+        Some(staker_setup.effective_block_state.number)
+    );
+    assert_eq!(staker.retired_balance, staker_setup.retired_stake);
+
+    let validator = staker_setup
+        .staking_contract
+        .get_validator(&data_store.read(&db_txn), &staker_setup.validator_address)
+        .unwrap();
+    assert_eq!(validator.num_stakers, 1);
+    assert_eq!(
+        validator.total_stake,
+        Coin::from_u64_unchecked(Policy::VALIDATOR_DEPOSIT)
+    );
+}
+
+/// Adding stake cannot violate minimum stake for non-retired balances.
+#[test]
+fn add_stake_enforces_minimum_stake_works() {
+    // -----------------------------------
+    // Test setup:
+    // -----------------------------------
+    let mut staker_setup = StakerSetup::setup_staker_with_inactive_retired_balance(
+        ValidatorState::Active,
+        0,
+        0,
+        50_000_000,
+        Policy::max_supported_version(),
+    );
+    let data_store = staker_setup
+        .accounts
+        .data_store(&Policy::STAKING_CONTRACT_ADDRESS);
+    let mut db_txn = staker_setup.env.write_transaction();
+    let mut db_txn = (&mut db_txn).into();
+    let staker_keypair = ed25519_key_pair(STAKER_PRIVATE_KEY);
+
+    // -----------------------------------
+    // Test execution:
+    // -----------------------------------
+    // Cannot add less than minimum stake.
+    let tx = make_signed_incoming_transaction(
+        IncomingStakingTransactionData::AddStake {
+            staker_address: staker_setup.staker_address.clone(),
+        },
+        Policy::MINIMUM_STAKE - 1,
+        &staker_keypair,
+    );
+    assert_eq!(
+        tx.verify(NetworkId::UnitAlbatross),
+        Err(TransactionError::InvalidValue)
+    );
+
+    // Can add in the valid case.
+    let tx = make_signed_incoming_transaction(
+        IncomingStakingTransactionData::AddStake {
+            staker_address: staker_setup.staker_address.clone(),
+        },
+        Policy::MINIMUM_STAKE,
+        &staker_keypair,
+    );
+
+    let mut tx_logs = TransactionLog::empty();
+    let receipt = staker_setup
+        .staking_contract
+        .commit_incoming_transaction(
+            &tx,
+            &staker_setup.before_release_block_state,
+            data_store.write(&mut db_txn),
+            &mut tx_logs,
+        )
+        .expect("Failed to commit transaction");
+
+    assert_eq!(
+        receipt,
+        Some(
+            AddStakeReceipt {
+                credited_balance: BalanceType::Retired
+            }
+            .into()
+        )
+    );
+
+    assert_eq!(
+        tx_logs.logs,
+        vec![Log::Stake {
+            staker_address: staker_setup.staker_address.clone(),
+            validator_address: Some(staker_setup.validator_address.clone()),
+            value: Coin::from_u64_unchecked(Policy::MINIMUM_STAKE),
+            credited_balance: BalanceType::Retired,
+        }]
+    );
+
+    let staker = staker_setup
+        .staking_contract
+        .get_staker(&data_store.read(&db_txn), &staker_setup.staker_address)
+        .expect("Staker should exist");
+
+    assert_eq!(
+        staker.retired_balance,
+        Coin::from_u64_unchecked(Policy::MINIMUM_STAKE + 50_000_000),
+    );
+    assert_eq!(staker.active_balance, Coin::ZERO,);
+    assert_eq!(staker.inactive_balance, Coin::ZERO);
+    assert_eq!(staker.inactive_from, None);
+
+    let validator = staker_setup
+        .staking_contract
+        .get_validator(&data_store.read(&db_txn), &staker_setup.validator_address)
+        .unwrap();
+
+    assert_eq!(validator.num_stakers, 1);
+    assert_eq!(
+        validator.total_stake,
+        Coin::from_u64_unchecked(Policy::VALIDATOR_DEPOSIT)
+    );
+}
+
+/* Legacy Add Stake testing */
+
+#[test]
+fn add_stake_legacy_works() {
+    add_stake(Policy::ADD_STAKE_PROTOCOL_UPGRADE_VERSION - 1);
+}
+
+#[test]
+fn add_stake_priority_legacy_works() {
+    // -----------------------------------
+    // Test setup:
+    // -----------------------------------
+    let mut staker_setup = StakerSetup::setup_staker_with_inactive_retired_balance(
+        ValidatorState::Active,
+        0,
+        50_000_000,
+        Policy::MINIMUM_STAKE + 1,
+        Policy::ADD_STAKE_PROTOCOL_UPGRADE_VERSION - 1,
+    );
+    assert!(staker_setup.active_stake < staker_setup.retired_stake);
+    assert!(staker_setup.active_stake < staker_setup.inactive_stake);
+    let data_store = staker_setup
+        .accounts
+        .data_store(&Policy::STAKING_CONTRACT_ADDRESS);
+    let mut db_txn = staker_setup.env.write_transaction();
+    let mut db_txn = (&mut db_txn).into();
+    let staker_keypair = ed25519_key_pair(STAKER_PRIVATE_KEY);
+
+    // -----------------------------------
+    // Test execution:
+    // -----------------------------------
+    // Add stake operation credits to staker.
+    let tx = make_signed_incoming_transaction(
+        IncomingStakingTransactionData::AddStake {
+            staker_address: staker_setup.staker_address.clone(),
+        },
+        Policy::MINIMUM_STAKE,
+        &staker_keypair,
+    );
+
+    let mut tx_logs = TransactionLog::empty();
+    let receipt = staker_setup
+        .staking_contract
+        .commit_incoming_transaction(
+            &tx,
+            &staker_setup.before_release_block_state,
+            data_store.write(&mut db_txn),
+            &mut tx_logs,
+        )
+        .expect("Failed to commit transaction");
+
+    assert_eq!(
+        receipt,
+        Some(
+            AddStakeReceipt {
+                credited_balance: BalanceType::Active
+            }
+            .into()
+        )
+    );
+
+    assert_eq!(
+        tx_logs.logs,
+        vec![Log::Stake {
+            staker_address: staker_setup.staker_address.clone(),
+            validator_address: Some(staker_setup.validator_address.clone()),
+            value: Coin::from_u64_unchecked(Policy::MINIMUM_STAKE),
+            credited_balance: BalanceType::Active,
+        }]
+    );
+
+    let staker = staker_setup
+        .staking_contract
+        .get_staker(&data_store.read(&db_txn), &staker_setup.staker_address)
+        .expect("Staker should exist");
+    assert_eq!(
+        staker.delegation,
+        Some(staker_setup.validator_address.clone())
+    );
+
+    assert_eq!(
+        staker.active_balance,
+        Coin::from_u64_unchecked(Policy::MINIMUM_STAKE)
+    );
+    assert_eq!(
+        staker.inactive_balance,
+        Coin::from_u64_unchecked(Policy::MINIMUM_STAKE * 5)
+    );
+    assert_eq!(staker.inactive_from, Some(328));
+    assert_eq!(
+        staker.retired_balance,
+        Coin::from_u64_unchecked(Policy::MINIMUM_STAKE + 1)
+    );
+
+    let validator = staker_setup
+        .staking_contract
+        .get_validator(&data_store.read(&db_txn), &staker_setup.validator_address)
+        .unwrap();
+
+    assert_eq!(validator.num_stakers, 1);
+    assert_eq!(
+        validator.total_stake,
+        Coin::from_u64_unchecked(Policy::VALIDATOR_DEPOSIT + Policy::MINIMUM_STAKE)
+    );
+
+    // Reverts correctly.
+    staker_setup
+        .staking_contract
+        .revert_incoming_transaction(
+            &tx,
+            &staker_setup.before_release_block_state,
+            receipt,
+            data_store.write(&mut db_txn),
+            &mut TransactionLog::empty(),
+        )
+        .expect("Failed to commit transaction");
+
+    let staker = staker_setup
+        .staking_contract
+        .get_staker(&data_store.read(&db_txn), &staker_setup.staker_address)
+        .expect("Staker should exist");
+    assert_eq!(
+        staker.delegation,
+        Some(staker_setup.validator_address.clone())
+    );
+
+    assert_eq!(staker.active_balance, staker_setup.active_stake);
+    assert_eq!(staker.inactive_balance, staker_setup.inactive_stake);
+    assert_eq!(
+        staker.inactive_from,
+        Some(staker_setup.effective_block_state.number)
+    );
+    assert_eq!(staker.retired_balance, staker_setup.retired_stake);
+
+    let validator = staker_setup
+        .staking_contract
+        .get_validator(&data_store.read(&db_txn), &staker_setup.validator_address)
+        .unwrap();
+    assert_eq!(validator.num_stakers, 1);
+    assert_eq!(
+        validator.total_stake,
+        Coin::from_u64_unchecked(Policy::VALIDATOR_DEPOSIT)
+    );
+}
+
+#[test]
+fn add_stake_enforces_minimum_stake_legacy_works() {
+    // -----------------------------------
+    // Test setup:
+    // -----------------------------------
+    let mut staker_setup = StakerSetup::setup_staker_with_inactive_retired_balance(
+        ValidatorState::Active,
+        0,
+        0,
+        50_000_000,
+        Policy::ADD_STAKE_PROTOCOL_UPGRADE_VERSION - 1,
+    );
+    let data_store = staker_setup
+        .accounts
+        .data_store(&Policy::STAKING_CONTRACT_ADDRESS);
+    let mut db_txn = staker_setup.env.write_transaction();
+    let mut db_txn = (&mut db_txn).into();
+    let staker_keypair = ed25519_key_pair(STAKER_PRIVATE_KEY);
+
+    // -----------------------------------
+    // Test execution:
+    // -----------------------------------
+    // Cannot add less than minimum stake.
+    let tx = make_signed_incoming_transaction(
+        IncomingStakingTransactionData::AddStake {
+            staker_address: staker_setup.staker_address.clone(),
+        },
+        Policy::MINIMUM_STAKE - 1,
+        &staker_keypair,
+    );
+    assert_eq!(
+        tx.verify(NetworkId::UnitAlbatross),
+        Err(TransactionError::InvalidValue)
+    );
+
+    // Can add in the valid case.
+    let tx = make_signed_incoming_transaction(
+        IncomingStakingTransactionData::AddStake {
+            staker_address: staker_setup.staker_address.clone(),
+        },
+        Policy::MINIMUM_STAKE,
+        &staker_keypair,
+    );
+
+    let mut tx_logs = TransactionLog::empty();
+    let receipt = staker_setup
+        .staking_contract
+        .commit_incoming_transaction(
+            &tx,
+            &staker_setup.before_release_block_state,
+            data_store.write(&mut db_txn),
+            &mut tx_logs,
+        )
+        .expect("Failed to commit transaction");
+
+    assert_eq!(
+        receipt,
+        Some(
+            AddStakeReceipt {
+                credited_balance: BalanceType::Active
+            }
+            .into()
+        )
+    );
+
+    assert_eq!(
+        tx_logs.logs,
+        vec![Log::Stake {
+            staker_address: staker_setup.staker_address.clone(),
+            validator_address: Some(staker_setup.validator_address.clone()),
+            value: Coin::from_u64_unchecked(Policy::MINIMUM_STAKE),
+            credited_balance: BalanceType::Active,
+        }]
+    );
+
+    let staker = staker_setup
+        .staking_contract
+        .get_staker(&data_store.read(&db_txn), &staker_setup.staker_address)
+        .expect("Staker should exist");
+
+    assert_eq!(
+        staker.active_balance,
+        Coin::from_u64_unchecked(Policy::MINIMUM_STAKE),
+    );
+    assert_eq!(staker.inactive_balance, Coin::ZERO);
+    assert_eq!(staker.inactive_from, None);
+    assert_eq!(staker.retired_balance, Coin::from_u64_unchecked(50_000_000));
+
+    let validator = staker_setup
+        .staking_contract
+        .get_validator(&data_store.read(&db_txn), &staker_setup.validator_address)
+        .unwrap();
+
+    assert_eq!(validator.num_stakers, 1);
+    assert_eq!(
+        validator.total_stake,
+        Coin::from_u64_unchecked(Policy::VALIDATOR_DEPOSIT + Policy::MINIMUM_STAKE)
+    );
+}
