+++ primitives/account/tests/staking_contract/staker.rs
@@ -1,9 +1,7 @@
 use nimiq_account::*;
 use nimiq_database::{mdbx::MdbxDatabase, traits::Database};
 use nimiq_keys::Address;
-use nimiq_primitives::{
-    account::AccountError, coin::Coin, policy::Policy, transaction::TransactionError,
-};
+use nimiq_primitives::{account::AccountError, coin::Coin, policy::Policy};
 use nimiq_test_log::test;
 use nimiq_transaction::{
     account::staking_contract::{IncomingStakingTransactionData, OutgoingStakingTransactionData},
@@ -86,6 +84,7 @@ fn prepare_second_validator_for_redelegation(
         active_stake,
         inactive_stake,
         retired_stake,
+        Policy::max_supported_version(),
     );
     let data_store = staker_setup
         .accounts
@@ -310,651 +309,6 @@ fn create_staker_works() {
     );
 }
 
-#[test]
-fn add_stake_works() {
-    let env = MdbxDatabase::new_volatile(Default::default()).unwrap();
-    let accounts = Accounts::new(env.clone());
-    let data_store = accounts.data_store(&Policy::STAKING_CONTRACT_ADDRESS);
-    let block_state = BlockState::new(2, 2, Policy::max_supported_version());
-    let mut db_txn = env.write_transaction();
-    let mut db_txn = (&mut db_txn).into();
-
-    let (validator_address, staker_address, mut staking_contract) =
-        make_sample_contract(data_store.write(&mut db_txn), Some(150_000_000));
-    let staker_address = staker_address.unwrap();
-    let staker_keypair = ed25519_key_pair(STAKER_PRIVATE_KEY);
-
-    // Works in the valid case.
-    let tx = make_signed_incoming_transaction(
-        IncomingStakingTransactionData::AddStake {
-            staker_address: staker_address.clone(),
-        },
-        150_000_000,
-        &staker_keypair,
-    );
-
-    let mut tx_logger = TransactionLog::empty();
-    let receipt = staking_contract
-        .commit_incoming_transaction(
-            &tx,
-            &block_state,
-            data_store.write(&mut db_txn),
-            &mut tx_logger,
-        )
-        .expect("Failed to commit transaction");
-
-    assert_eq!(
-        receipt,
-        Some(
-            AddStakeReceipt {
-                credited_balance: BalanceType::Active
-            }
-            .into()
-        )
-    );
-    assert_eq!(
-        tx_logger.logs,
-        vec![Log::Stake {
-            staker_address: staker_address.clone(),
-            validator_address: Some(validator_address.clone()),
-            value: tx.value,
-            credited_balance: BalanceType::Active,
-        }]
-    );
-
-    let staker = staking_contract
-        .get_staker(&data_store.read(&db_txn), &staker_address)
-        .expect("Staker should exist");
-
-    assert_eq!(staker.address, staker_address);
-    assert_eq!(staker.active_balance, Coin::from_u64_unchecked(300_000_000));
-    assert_eq!(staker.delegation, Some(validator_address.clone()));
-
-    let validator = staking_contract
-        .get_validator(&data_store.read(&db_txn), &validator_address)
-        .expect("Validator should exist");
-
-    assert_eq!(
-        validator.total_stake,
-        Coin::from_u64_unchecked(Policy::VALIDATOR_DEPOSIT + 300_000_000)
-    );
-    assert_eq!(validator.num_stakers, 1);
-
-    assert_eq!(
-        staking_contract.balance,
-        Coin::from_u64_unchecked(Policy::VALIDATOR_DEPOSIT + 300_000_000)
-    );
-
-    assert_eq!(
-        staking_contract.active_validators.get(&validator_address),
-        Some(&Coin::from_u64_unchecked(
-            Policy::VALIDATOR_DEPOSIT + 300_000_000
-        ))
-    );
-
-    // Revert the transaction.
-    let mut tx_logger = TransactionLog::empty();
-    staking_contract
-        .revert_incoming_transaction(
-            &tx,
-            &block_state,
-            receipt,
-            data_store.write(&mut db_txn),
-            &mut tx_logger,
-        )
-        .expect("Failed to revert transaction");
-
-    assert_eq!(
-        tx_logger.logs,
-        vec![Log::Stake {
-            staker_address: staker_address.clone(),
-            validator_address: Some(validator_address.clone()),
-            value: tx.value,
-            credited_balance: BalanceType::Active,
-        }]
-    );
-
-    let staker = staking_contract
-        .get_staker(&data_store.read(&db_txn), &staker_address)
-        .expect("Staker should exist");
-
-    assert_eq!(staker.address, staker_address);
-    assert_eq!(staker.active_balance, Coin::from_u64_unchecked(150_000_000));
-    assert_eq!(staker.delegation, Some(validator_address.clone()));
-
-    let validator = staking_contract
-        .get_validator(&data_store.read(&db_txn), &validator_address)
-        .expect("Validator should exist");
-
-    assert_eq!(
-        validator.total_stake,
-        Coin::from_u64_unchecked(Policy::VALIDATOR_DEPOSIT + 150_000_000)
-    );
-    assert_eq!(validator.num_stakers, 1);
-
-    assert_eq!(
-        staking_contract.balance,
-        Coin::from_u64_unchecked(Policy::VALIDATOR_DEPOSIT + 150_000_000)
-    );
-
-    assert_eq!(
-        staking_contract.active_validators.get(&validator_address),
-        Some(&Coin::from_u64_unchecked(
-            Policy::VALIDATOR_DEPOSIT + 150_000_000
-        ))
-    );
-}
-
-/// Adding stake should give priority to active stake if any.
-/// , otherwise it
-/// should credit the biggest balance between inactive and retired.
-#[test]
-fn add_stake_policy_priority_to_active_balance() {
-    // -----------------------------------
-    // Test setup:
-    // -----------------------------------
-    let mut staker_setup = StakerSetup::setup_staker_with_inactive_retired_balance(
-        ValidatorState::Active,
-        Policy::MINIMUM_STAKE,
-        Policy::MINIMUM_STAKE + 1,
-        50_000_000,
-    );
-    assert!(staker_setup.active_stake < staker_setup.retired_stake);
-    assert!(staker_setup.active_stake < staker_setup.inactive_stake);
-    let data_store = staker_setup
-        .accounts
-        .data_store(&Policy::STAKING_CONTRACT_ADDRESS);
-    let mut db_txn = staker_setup.env.write_transaction();
-    let mut db_txn = (&mut db_txn).into();
-    let staker_keypair = ed25519_key_pair(STAKER_PRIVATE_KEY);
-
-    // -----------------------------------
-    // Test execution:
-    // -----------------------------------
-    // Add stake operation credits to active stake.
-    let tx = make_signed_incoming_transaction(
-        IncomingStakingTransactionData::AddStake {
-            staker_address: staker_setup.staker_address.clone(),
-        },
-        Policy::MINIMUM_STAKE,
-        &staker_keypair,
-    );
-
-    let mut tx_logs = TransactionLog::empty();
-    let receipt = staker_setup
-        .staking_contract
-        .commit_incoming_transaction(
-            &tx,
-            &staker_setup.before_release_block_state,
-            data_store.write(&mut db_txn),
-            &mut tx_logs,
-        )
-        .expect("Failed to commit transaction");
-
-    assert_eq!(
-        receipt,
-        Some(
-            AddStakeReceipt {
-                credited_balance: BalanceType::Active
-            }
-            .into()
-        )
-    );
-
-    assert_eq!(
-        tx_logs.logs,
-        vec![Log::Stake {
-            staker_address: staker_setup.staker_address.clone(),
-            validator_address: Some(staker_setup.validator_address.clone()),
-            value: Coin::from_u64_unchecked(Policy::MINIMUM_STAKE),
-            credited_balance: BalanceType::Active,
-        }]
-    );
-
-    let staker = staker_setup
-        .staking_contract
-        .get_staker(&data_store.read(&db_txn), &staker_setup.staker_address)
-        .expect("Staker should exist");
-
-    assert_eq!(
-        staker.delegation,
-        Some(staker_setup.validator_address.clone())
-    );
-    assert_eq!(
-        staker.active_balance,
-        Coin::from_u64_unchecked(Policy::MINIMUM_STAKE * 2)
-    );
-    assert_eq!(
-        staker.inactive_balance,
-        Coin::from_u64_unchecked(Policy::MINIMUM_STAKE + 1)
-    );
-    assert_eq!(staker.inactive_from, Some(328));
-    assert_eq!(
-        staker.retired_balance,
-        Coin::from_u64_unchecked(Policy::MINIMUM_STAKE * 5)
-    );
-
-    let validator = staker_setup
-        .staking_contract
-        .get_validator(&data_store.read(&db_txn), &staker_setup.validator_address)
-        .unwrap();
-
-    assert_eq!(validator.num_stakers, 1);
-    assert_eq!(
-        validator.total_stake,
-        Coin::from_u64_unchecked(Policy::VALIDATOR_DEPOSIT + Policy::MINIMUM_STAKE * 2)
-    );
-
-    // Reverts correctly.
-    staker_setup
-        .staking_contract
-        .revert_incoming_transaction(
-            &tx,
-            &staker_setup.before_release_block_state,
-            receipt,
-            data_store.write(&mut db_txn),
-            &mut TransactionLog::empty(),
-        )
-        .expect("Failed to commit transaction");
-
-    let staker = staker_setup
-        .staking_contract
-        .get_staker(&data_store.read(&db_txn), &staker_setup.staker_address)
-        .expect("Staker should exist");
-    assert_eq!(
-        staker.delegation,
-        Some(staker_setup.validator_address.clone())
-    );
-
-    assert_eq!(staker.active_balance, staker_setup.active_stake);
-    assert_eq!(staker.inactive_balance, staker_setup.inactive_stake);
-    assert_eq!(
-        staker.inactive_from,
-        Some(staker_setup.effective_block_state.number)
-    );
-    assert_eq!(staker.retired_balance, staker_setup.retired_stake);
-
-    let validator = staker_setup
-        .staking_contract
-        .get_validator(&data_store.read(&db_txn), &staker_setup.validator_address)
-        .unwrap();
-    assert_eq!(validator.num_stakers, 1);
-    assert_eq!(
-        validator.total_stake,
-        Coin::from_u64_unchecked(Policy::VALIDATOR_DEPOSIT + Policy::MINIMUM_STAKE)
-    );
-}
-
-/// Adding stake in absence of active balance should prioritize the maximum between the inactive and retired balance.
-#[test]
-fn add_stake_policy_maximum_between_inactive_retired() {
-    // -----------------------------------
-    // Test setup:
-    // -----------------------------------
-    let mut staker_setup = StakerSetup::setup_staker_with_inactive_retired_balance(
-        ValidatorState::Active,
-        0,
-        Policy::MINIMUM_STAKE + 1,
-        50_000_000,
-    );
-    assert!(staker_setup.active_stake < staker_setup.retired_stake);
-    assert!(staker_setup.active_stake < staker_setup.inactive_stake);
-    let data_store = staker_setup
-        .accounts
-        .data_store(&Policy::STAKING_CONTRACT_ADDRESS);
-    let mut db_txn = staker_setup.env.write_transaction();
-    let mut db_txn = (&mut db_txn).into();
-    let staker_keypair = ed25519_key_pair(STAKER_PRIVATE_KEY);
-
-    // -----------------------------------
-    // Test execution:
-    // -----------------------------------
-    // Add stake operation credits to active stake.
-    let tx = make_signed_incoming_transaction(
-        IncomingStakingTransactionData::AddStake {
-            staker_address: staker_setup.staker_address.clone(),
-        },
-        Policy::MINIMUM_STAKE,
-        &staker_keypair,
-    );
-
-    let mut tx_logs = TransactionLog::empty();
-    let receipt = staker_setup
-        .staking_contract
-        .commit_incoming_transaction(
-            &tx,
-            &staker_setup.before_release_block_state,
-            data_store.write(&mut db_txn),
-            &mut tx_logs,
-        )
-        .expect("Failed to commit transaction");
-
-    assert_eq!(
-        receipt,
-        Some(
-            AddStakeReceipt {
-                credited_balance: BalanceType::Retired
-            }
-            .into()
-        )
-    );
-
-    assert_eq!(
-        tx_logs.logs,
-        vec![Log::Stake {
-            staker_address: staker_setup.staker_address.clone(),
-            validator_address: Some(staker_setup.validator_address.clone()),
-            value: Coin::from_u64_unchecked(Policy::MINIMUM_STAKE),
-            credited_balance: BalanceType::Retired,
-        }]
-    );
-
-    let staker = staker_setup
-        .staking_contract
-        .get_staker(&data_store.read(&db_txn), &staker_setup.staker_address)
-        .expect("Staker should exist");
-
-    assert_eq!(
-        staker.delegation,
-        Some(staker_setup.validator_address.clone())
-    );
-    assert_eq!(staker.active_balance, Coin::ZERO);
-    assert_eq!(
-        staker.inactive_balance,
-        Coin::from_u64_unchecked(Policy::MINIMUM_STAKE + 1)
-    );
-    assert_eq!(staker.inactive_from, Some(328));
-    assert_eq!(
-        staker.retired_balance,
-        Coin::from_u64_unchecked(Policy::MINIMUM_STAKE * 6)
-    );
-
-    let validator = staker_setup
-        .staking_contract
-        .get_validator(&data_store.read(&db_txn), &staker_setup.validator_address)
-        .unwrap();
-
-    assert_eq!(validator.num_stakers, 1);
-    assert_eq!(
-        validator.total_stake,
-        Coin::from_u64_unchecked(Policy::VALIDATOR_DEPOSIT)
-    );
-
-    // Reverts correctly.
-    staker_setup
-        .staking_contract
-        .revert_incoming_transaction(
-            &tx,
-            &staker_setup.before_release_block_state,
-            receipt,
-            data_store.write(&mut db_txn),
-            &mut TransactionLog::empty(),
-        )
-        .expect("Failed to commit transaction");
-
-    let staker = staker_setup
-        .staking_contract
-        .get_staker(&data_store.read(&db_txn), &staker_setup.staker_address)
-        .expect("Staker should exist");
-    assert_eq!(
-        staker.delegation,
-        Some(staker_setup.validator_address.clone())
-    );
-
-    assert_eq!(staker.active_balance, staker_setup.active_stake);
-    assert_eq!(staker.inactive_balance, staker_setup.inactive_stake);
-    assert_eq!(
-        staker.inactive_from,
-        Some(staker_setup.effective_block_state.number)
-    );
-    assert_eq!(staker.retired_balance, staker_setup.retired_stake);
-
-    let validator = staker_setup
-        .staking_contract
-        .get_validator(&data_store.read(&db_txn), &staker_setup.validator_address)
-        .unwrap();
-    assert_eq!(validator.num_stakers, 1);
-    assert_eq!(
-        validator.total_stake,
-        Coin::from_u64_unchecked(Policy::VALIDATOR_DEPOSIT)
-    );
-}
-
-/// Adding stake to the inactive stake works.
-#[test]
-fn add_stake_policy_to_inactive_works() {
-    // -----------------------------------
-    // Test setup:
-    // -----------------------------------
-    let mut staker_setup = StakerSetup::setup_staker_with_inactive_retired_balance(
-        ValidatorState::Active,
-        0,
-        50_000_000,
-        Policy::MINIMUM_STAKE + 1,
-    );
-    assert!(staker_setup.active_stake < staker_setup.retired_stake);
-    assert!(staker_setup.active_stake < staker_setup.inactive_stake);
-    let data_store = staker_setup
-        .accounts
-        .data_store(&Policy::STAKING_CONTRACT_ADDRESS);
-    let mut db_txn = staker_setup.env.write_transaction();
-    let mut db_txn = (&mut db_txn).into();
-    let staker_keypair = ed25519_key_pair(STAKER_PRIVATE_KEY);
-
-    // -----------------------------------
-    // Test execution:
-    // -----------------------------------
-    // Add stake operation credits to active stake.
-    let tx = make_signed_incoming_transaction(
-        IncomingStakingTransactionData::AddStake {
-            staker_address: staker_setup.staker_address.clone(),
-        },
-        Policy::MINIMUM_STAKE,
-        &staker_keypair,
-    );
-
-    let mut tx_logs = TransactionLog::empty();
-    let receipt = staker_setup
-        .staking_contract
-        .commit_incoming_transaction(
-            &tx,
-            &staker_setup.before_release_block_state,
-            data_store.write(&mut db_txn),
-            &mut tx_logs,
-        )
-        .expect("Failed to commit transaction");
-
-    assert_eq!(
-        receipt,
-        Some(
-            AddStakeReceipt {
-                credited_balance: BalanceType::Inactive
-            }
-            .into()
-        )
-    );
-
-    assert_eq!(
-        tx_logs.logs,
-        vec![Log::Stake {
-            staker_address: staker_setup.staker_address.clone(),
-            validator_address: Some(staker_setup.validator_address.clone()),
-            value: Coin::from_u64_unchecked(Policy::MINIMUM_STAKE),
-            credited_balance: BalanceType::Inactive,
-        }]
-    );
-
-    let staker = staker_setup
-        .staking_contract
-        .get_staker(&data_store.read(&db_txn), &staker_setup.staker_address)
-        .expect("Staker should exist");
-    assert_eq!(
-        staker.delegation,
-        Some(staker_setup.validator_address.clone())
-    );
-
-    assert_eq!(staker.active_balance, Coin::ZERO);
-    assert_eq!(
-        staker.inactive_balance,
-        Coin::from_u64_unchecked(Policy::MINIMUM_STAKE * 6)
-    );
-    assert_eq!(staker.inactive_from, Some(328));
-    assert_eq!(
-        staker.retired_balance,
-        Coin::from_u64_unchecked(Policy::MINIMUM_STAKE + 1)
-    );
-
-    let validator = staker_setup
-        .staking_contract
-        .get_validator(&data_store.read(&db_txn), &staker_setup.validator_address)
-        .unwrap();
-
-    assert_eq!(validator.num_stakers, 1);
-    assert_eq!(
-        validator.total_stake,
-        Coin::from_u64_unchecked(Policy::VALIDATOR_DEPOSIT)
-    );
-
-    // Reverts correctly.
-    staker_setup
-        .staking_contract
-        .revert_incoming_transaction(
-            &tx,
-            &staker_setup.before_release_block_state,
-            receipt,
-            data_store.write(&mut db_txn),
-            &mut TransactionLog::empty(),
-        )
-        .expect("Failed to commit transaction");
-
-    let staker = staker_setup
-        .staking_contract
-        .get_staker(&data_store.read(&db_txn), &staker_setup.staker_address)
-        .expect("Staker should exist");
-    assert_eq!(
-        staker.delegation,
-        Some(staker_setup.validator_address.clone())
-    );
-
-    assert_eq!(staker.active_balance, staker_setup.active_stake);
-    assert_eq!(staker.inactive_balance, staker_setup.inactive_stake);
-    assert_eq!(
-        staker.inactive_from,
-        Some(staker_setup.effective_block_state.number)
-    );
-    assert_eq!(staker.retired_balance, staker_setup.retired_stake);
-
-    let validator = staker_setup
-        .staking_contract
-        .get_validator(&data_store.read(&db_txn), &staker_setup.validator_address)
-        .unwrap();
-    assert_eq!(validator.num_stakers, 1);
-    assert_eq!(
-        validator.total_stake,
-        Coin::from_u64_unchecked(Policy::VALIDATOR_DEPOSIT)
-    );
-}
-
-/// Adding stake cannot violate minimum stake for non-retired balances.
-#[test]
-fn add_stake_enforces_minimum_stake() {
-    // -----------------------------------
-    // Test setup:
-    // -----------------------------------
-    let mut staker_setup = StakerSetup::setup_staker_with_inactive_retired_balance(
-        ValidatorState::Active,
-        0,
-        0,
-        50_000_000,
-    );
-    let data_store = staker_setup
-        .accounts
-        .data_store(&Policy::STAKING_CONTRACT_ADDRESS);
-    let mut db_txn = staker_setup.env.write_transaction();
-    let mut db_txn = (&mut db_txn).into();
-    let staker_keypair = ed25519_key_pair(STAKER_PRIVATE_KEY);
-
-    // -----------------------------------
-    // Test execution:
-    // -----------------------------------
-    // Cannot add less than minimum stake.
-    let tx = make_signed_incoming_transaction(
-        IncomingStakingTransactionData::AddStake {
-            staker_address: staker_setup.staker_address.clone(),
-        },
-        Policy::MINIMUM_STAKE - 1,
-        &staker_keypair,
-    );
-    assert_eq!(
-        tx.verify(NetworkId::UnitAlbatross),
-        Err(TransactionError::InvalidValue)
-    );
-
-    // Can add in the valid case.
-    let tx = make_signed_incoming_transaction(
-        IncomingStakingTransactionData::AddStake {
-            staker_address: staker_setup.staker_address.clone(),
-        },
-        Policy::MINIMUM_STAKE,
-        &staker_keypair,
-    );
-
-    let mut tx_logs = TransactionLog::empty();
-    let receipt = staker_setup
-        .staking_contract
-        .commit_incoming_transaction(
-            &tx,
-            &staker_setup.before_release_block_state,
-            data_store.write(&mut db_txn),
-            &mut tx_logs,
-        )
-        .expect("Failed to commit transaction");
-
-    assert_eq!(
-        receipt,
-        Some(
-            AddStakeReceipt {
-                credited_balance: BalanceType::Retired
-            }
-            .into()
-        )
-    );
-
-    assert_eq!(
-        tx_logs.logs,
-        vec![Log::Stake {
-            staker_address: staker_setup.staker_address.clone(),
-            validator_address: Some(staker_setup.validator_address.clone()),
-            value: Coin::from_u64_unchecked(Policy::MINIMUM_STAKE),
-            credited_balance: BalanceType::Retired,
-        }]
-    );
-
-    let staker = staker_setup
-        .staking_contract
-        .get_staker(&data_store.read(&db_txn), &staker_setup.staker_address)
-        .expect("Staker should exist");
-
-    assert_eq!(
-        staker.retired_balance,
-        Coin::from_u64_unchecked(Policy::MINIMUM_STAKE + 50_000_000),
-    );
-    assert_eq!(staker.active_balance, Coin::ZERO,);
-    assert_eq!(staker.inactive_balance, Coin::ZERO);
-    assert_eq!(staker.inactive_from, None);
-
-    let validator = staker_setup
-        .staking_contract
-        .get_validator(&data_store.read(&db_txn), &staker_setup.validator_address)
-        .unwrap();
-
-    assert_eq!(validator.num_stakers, 1);
-    assert_eq!(
-        validator.total_stake,
-        Coin::from_u64_unchecked(Policy::VALIDATOR_DEPOSIT)
-    );
-}
-
 /// Updating inactive balance resets counter
 #[test]
 fn can_set_inactive_balance() {
@@ -966,6 +320,7 @@ fn can_set_inactive_balance() {
         50_000_000,
         50_000_000,
         10_000_000,
+        Policy::max_supported_version(),
     );
     let data_store = staker_setup
         .accounts
@@ -1164,6 +519,7 @@ fn retire_inactive_stake_works_with_min_active_stake() {
         Policy::MINIMUM_STAKE,
         Policy::MINIMUM_STAKE + 1,
         1,
+        Policy::max_supported_version(),
     );
     let data_store = staker_setup
         .accounts
@@ -1394,6 +750,7 @@ fn retire_stake_does_not_violate_jail_or_inactive_releases() {
         0,
         Policy::MINIMUM_STAKE,
         10_000_000,
+        Policy::max_supported_version(),
     );
     let data_store = staker_setup
         .accounts
@@ -1427,6 +784,7 @@ fn retire_stake_does_not_violate_jail_or_inactive_releases() {
         0,
         Policy::MINIMUM_STAKE,
         20_000_000,
+        Policy::max_supported_version(),
     );
     let data_store = staker_setup
         .accounts
@@ -1486,6 +844,7 @@ fn retire_stake_does_not_violate_minimum_active_stake() {
         0,
         Policy::MINIMUM_STAKE + 1,
         1,
+        Policy::max_supported_version(),
     );
     let data_store = staker_setup
         .accounts
@@ -1592,6 +951,7 @@ fn cannot_retire_active_stake() {
         Policy::MINIMUM_STAKE,
         0,
         1,
+        Policy::max_supported_version(),
     );
     let data_store = staker_setup
         .accounts
@@ -1946,6 +1306,7 @@ fn update_staker_with_stake_reactivation_works() {
         0,
         150_000_000,
         100_000_000,
+        Policy::max_supported_version(),
     );
     let data_store = staker_setup
         .accounts
@@ -2174,6 +1535,7 @@ fn update_staker_remove_delegation_with_stake_reactivation_works() {
         0,
         150_000_000,
         100_000_000,
+        Policy::max_supported_version(),
     );
     let data_store = staker_setup
         .accounts
@@ -2340,6 +1702,7 @@ fn update_staker_same_validator() {
         0,
         150_000_000,
         100_000_000,
+        Policy::max_supported_version(),
     );
     let data_store = staker_setup
         .accounts
@@ -2526,6 +1889,7 @@ fn remove_stake_works() {
         0,
         0,
         Policy::MINIMUM_STAKE * 2,
+        Policy::max_supported_version(),
     );
     let data_store = staker_setup
         .accounts
@@ -2739,6 +2103,7 @@ fn remove_stake_from_tombstone_works() {
         0,
         0,
         150_000_000,
+        Policy::max_supported_version(),
     );
     let data_store = staker_setup
         .accounts
@@ -2820,6 +2185,7 @@ fn can_only_remove_retired_balance() {
         50_000_000,
         50_000_000,
         50_000_000,
+        Policy::max_supported_version(),
     );
     let data_store = staker_setup
         .accounts
@@ -2972,7 +2338,7 @@ fn can_delegate_if_no_delegation_prior_delegation() {
     // Test setup:
     // -----------------------------------
     // Create a validator with no staker
-    let mut validator_setup = ValidatorSetup::new(None);
+    let mut validator_setup = ValidatorSetup::new(None, Policy::max_supported_version());
     let data_store = validator_setup
         .accounts
         .data_store(&Policy::STAKING_CONTRACT_ADDRESS);
@@ -3292,6 +2658,7 @@ fn can_reserve_and_release_balance() {
         40_000_000,
         10_000_000,
         60_000_000,
+        Policy::max_supported_version(),
     );
     let data_store = staker_setup
         .accounts
@@ -3379,6 +2746,7 @@ fn commit_failed_remove_stake_works_when_removing_staker() {
         0,
         0,
         Policy::MINIMUM_STAKE + 1,
+        Policy::max_supported_version(),
     );
     let data_store = staker_setup
         .accounts
@@ -3591,6 +2959,7 @@ fn commit_failed_remove_stake_works_with_non_retired_balance() {
         0,
         Policy::MINIMUM_STAKE,
         Policy::MINIMUM_STAKE,
+        Policy::max_supported_version(),
     );
     let data_store = staker_setup
         .accounts
